backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
google-generativeai
psycopg2-binary
pytest
requests
